refactor: 분리
