+++ src/components/common/Header/@setting/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useSettingModal } from "./useSettingModal";
